AM-2848: Retire obsolete controllers and associated tests after review
